fix: extract streamWriteTracker and split metrics tests to stay under 1500-line ceiling
- Extract streamWriteTracker struct and all methods into new file
  stream_write_tracker.go (106 lines)
- Move stream-quality metric tests from metrics_test.go to
  metrics_stream_quality_test.go (consolidating all stream tests there)
- Remove gateway_http.go and metrics_test.go from FROZEN table
  in tools/source-lines.mjs

gateway_http.go: 1506 -> 1405 lines
metrics_test.go: 1549 -> 969 lines
Both are now under the 1500-line default ceiling.
