+++ backend/internal/server/metrics_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"net/http"
@@ -967,582 +966,4 @@ func TestMetricsStreamAttemptDurationIncludesClientBackpressure(t *testing.T) {
 	if got := metricLineValue(t, overheadLine); got >= backpressure.Seconds()*0.6 {
 		t.Fatalf("overhead must not include client backpressure, got %v", got)
 	}
-}
-
-// A successful stream records time_to_first_byte_seconds exactly once.
-func TestMetricsStreamRecordsTimeToFirstByteOnce(t *testing.T) {
-	_, server, secret := newMetricsTestServer(t, false)
-	app := server.Handler()
-	if code := chatOnce(t, app, secret, true); code != http.StatusOK {
-		t.Fatalf("streaming chat failed: %d", code)
-	}
-
-	body := scrapeMetrics(t, app)
-	line := findMetricLine(body, "tokenhub_gateway_time_to_first_byte_seconds_count")
-	if line == "" {
-		t.Fatalf("missing time_to_first_byte count:\n%s", body)
-	}
-	if got := metricLineValue(t, line); got != 1 {
-		t.Fatalf("expected TTFB observed once, got %v", got)
-	}
-}
-
-// Non-streamed requests do not create a time_to_first_byte_seconds series.
-func TestMetricsNonStreamRecordsNoTimeToFirstByte(t *testing.T) {
-	_, server, secret := newMetricsTestServer(t, false)
-	app := server.Handler()
-	if code := chatOnce(t, app, secret, false); code != http.StatusOK {
-		t.Fatalf("chat failed: %d", code)
-	}
-
-	body := scrapeMetrics(t, app)
-	if strings.Contains(body, "tokenhub_gateway_time_to_first_byte_seconds{") {
-		t.Fatalf("non-stream request must not produce TTFB series:\n%s", body)
-	}
-}
-
-// After failover, TTFB is attributed to the serving provider and observed once.
-func TestMetricsFailoverObservesTimeToFirstByteOnce(t *testing.T) {
-	var secondaryHits int32
-	primary := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusServiceUnavailable)
-		_, _ = io.WriteString(w, `{"error":{"message":"upstream down"}}`)
-	}))
-	defer primary.Close()
-	secondary := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		atomic.AddInt32(&secondaryHits, 1)
-		writeChatStreamChunk(w, "recovered")
-	}))
-	defer secondary.Close()
-
-	server, secret := newMetricsFailoverGateway(t, primary.URL, secondary.URL)
-	resp := postStream(t, server.Handler(), "/v1/chat/completions", map[string]any{
-		"model":    "metrics-failover-model",
-		"messages": []map[string]any{{"role": "user", "content": "hi"}},
-		"stream":   true,
-	}, secret)
-	if resp.Code != http.StatusOK {
-		t.Fatalf("expected fallback to succeed, got %d: %s", resp.Code, resp.Body.String())
-	}
-
-	body := scrapeMetrics(t, server.Handler())
-	count := findMetricLine(body, "tokenhub_gateway_time_to_first_byte_seconds_count", `provider_type="openai_compatible"`)
-	if count == "" {
-		t.Fatalf("missing TTFB count:\n%s", body)
-	}
-	if got := metricLineValue(t, count); got != 1 {
-		t.Fatalf("expected TTFB observed once after failover, got %v", got)
-	}
-	// The metric carries model and provider_type only; the request_total series
-	// below confirms the logical request was attributed to the serving provider.
-	if got := testutil.ToFloat64(server.metrics.requests.WithLabelValues(
-		"metrics-failover-model", ProviderOpenAICompatible, "prv_metrics_1", "rsrc_metrics_1", "200", metricsLabelUnset, "true",
-	)); got != 1 {
-		t.Fatalf("expected one logical request on serving provider, got %v", got)
-	}
-}
-
-// A committed stream that then fails counts as one interruption.
-func TestMetricsCommittedStreamFailureCountsInterruption(t *testing.T) {
-	primary := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("content-type", "text/event-stream")
-		flusher, _ := w.(http.Flusher)
-		_, _ = io.WriteString(w, "data: {\"id\":\"chatcmpl\",\"choices\":[{\"delta\":{\"content\":\"partial\"}}]}\n\n")
-		if flusher != nil {
-			flusher.Flush()
-		}
-		if hijacker, ok := w.(http.Hijacker); ok {
-			conn, _, hijackErr := hijacker.Hijack()
-			if hijackErr == nil {
-				_ = conn.Close()
-				return
-			}
-		}
-	}))
-	defer primary.Close()
-	secondary := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		writeChatStreamChunk(w, "should-not-be-used")
-	}))
-	defer secondary.Close()
-
-	server, secret := newMetricsFailoverGateway(t, primary.URL, secondary.URL)
-	resp := postStream(t, server.Handler(), "/v1/chat/completions", map[string]any{
-		"model":    "metrics-failover-model",
-		"messages": []map[string]any{{"role": "user", "content": "hi"}},
-		"stream":   true,
-	}, secret)
-
-	if strings.Contains(resp.Body.String(), "should-not-be-used") {
-		t.Fatalf("client received a second upstream's stream: %q", resp.Body.String())
-	}
-
-	body := scrapeMetrics(t, server.Handler())
-	interruption := findMetricLine(body, "tokenhub_gateway_stream_interruptions_total", `provider_id="prv_metrics_0"`)
-	if interruption == "" {
-		t.Fatalf("missing interruption series:\n%s", body)
-	}
-	if metricLineValue(t, interruption) != 1 {
-		t.Fatalf("expected one interruption")
-	}
-	if strings.Contains(interruption, `error_code="none"`) {
-		t.Fatalf("interruption must carry a non-none error_code:\n%s", interruption)
-	}
-
-	// TTFB is still observed for the interrupted stream.
-	if findMetricLine(body, "tokenhub_gateway_time_to_first_byte_seconds_count") == "" {
-		t.Fatalf("TTFB must still be observed for an interrupted stream")
-	}
-}
-
-// A stream that fails before the first byte is not an interruption.
-func TestMetricsPreCommitStreamFailureIsNotAnInterruption(t *testing.T) {
-	primary := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusServiceUnavailable)
-		_, _ = io.WriteString(w, `{"error":{"message":"upstream down"}}`)
-	}))
-	defer primary.Close()
-	secondary := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		writeChatStreamChunk(w, "recovered")
-	}))
-	defer secondary.Close()
-
-	server, secret := newMetricsFailoverGateway(t, primary.URL, secondary.URL)
-	resp := postStream(t, server.Handler(), "/v1/chat/completions", map[string]any{
-		"model":    "metrics-failover-model",
-		"messages": []map[string]any{{"role": "user", "content": "hi"}},
-		"stream":   true,
-	}, secret)
-	if resp.Code != http.StatusOK {
-		t.Fatalf("expected fallback to succeed, got %d: %s", resp.Code, resp.Body.String())
-	}
-
-	body := scrapeMetrics(t, server.Handler())
-	if strings.Contains(body, "tokenhub_gateway_stream_interruptions_total{") {
-		t.Fatalf("pre-commit failure must not count as interruption:\n%s", body)
-	}
-}
-
-// A client-disconnected stream records an interruption with the error_code that
-// classifyStreamError assigns to context.Canceled.
-func TestMetricsClientDisconnectInterruptionKeepsItsCode(t *testing.T) {
-	// Derive the code through the real classification chain rather than hardcoding
-	// a string that could drift.
-	err := classifyStreamError(context.Background(), context.Canceled, true)
-	httpErr := AsHTTPError(err)
-	expectedCode := httpErr.Code
-	if expectedCode == "" {
-		t.Fatal("could not derive client disconnect error code")
-	}
-
-	primary := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("content-type", "text/event-stream")
-		flusher, _ := w.(http.Flusher)
-		_, _ = io.WriteString(w, "data: {\"id\":\"chatcmpl\",\"choices\":[{\"delta\":{\"content\":\"partial\"}}]}\n\n")
-		if flusher != nil {
-			flusher.Flush()
-		}
-		// Wait until the client gives up.
-		<-r.Context().Done()
-	}))
-	defer primary.Close()
-
-	server, secret := newMetricsFailoverGateway(t, primary.URL, "http://localhost:1")
-
-	ctx, cancel := context.WithCancel(context.Background())
-	body, _ := json.Marshal(map[string]any{
-		"model":    "metrics-failover-model",
-		"messages": []map[string]any{{"role": "user", "content": "hi"}},
-		"stream":   true,
-	})
-	req := httptest.NewRequest(http.MethodPost, "/v1/chat/completions", strings.NewReader(string(body))).WithContext(ctx)
-	req.Header.Set("content-type", "application/json")
-	req.Header.Set("authorization", "Bearer "+secret)
-	rec := httptest.NewRecorder()
-
-	// Cancel shortly after the first byte has been written.
-	go func() {
-		time.Sleep(50 * time.Millisecond)
-		cancel()
-	}()
-	server.Handler().ServeHTTP(rec, req)
-
-	scrapeBody := scrapeMetrics(t, server.Handler())
-	interruption := findMetricLine(scrapeBody, "tokenhub_gateway_stream_interruptions_total", `error_code="`+expectedCode+`"`)
-	if interruption == "" {
-		t.Fatalf("missing interruption with code %q:\n%s", expectedCode, scrapeBody)
-	}
-}
-
-// newMetricsCodexTestServer wires one Codex subscription model with metrics
-// enabled. upstream returns the response the mock subscription should serve.
-func newMetricsCodexTestServer(t *testing.T, upstream func(*http.Request) *http.Response) (*Server, string) {
-	t.Helper()
-	store := NewMemoryStore()
-	project := store.CreateProject(Project{Name: "metrics-codex", Status: StatusActive})
-	const model = "metrics-codex-model"
-	_, secret, err := store.CreateAPIKey(project.ID, APIKey{
-		Name:    "metrics-codex-key",
-		Allowed: []string{model},
-		Status:  StatusActive,
-	}, "thk_metrics_codex")
-	if err != nil {
-		t.Fatal(err)
-	}
-	provider := store.AddProvider(Provider{
-		ID:      "prv_metrics_codex",
-		Name:    "Metrics Codex",
-		Type:    ProviderOpenAICodex,
-		Status:  StatusActive,
-		Healthy: true,
-	})
-	_, err = store.AddProviderResource(ProviderResource{
-		ID:           "rsrc_metrics_codex",
-		ProviderID:   provider.ID,
-		Name:         "Metrics Codex Account",
-		ResourceType: ProviderResourceOpenAISubscription,
-		Status:       StatusActive,
-		Healthy:      true,
-		Options:      codexCapabilityOptionsForTest(model),
-		Credentials: &ProviderResourceCredentials{
-			AccessToken: "access_metrics",
-			AccountID:   "account_metrics",
-		},
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	store.AddModel(Model{Name: model, Modality: "chat", Status: StatusActive})
-	store.AddRoute(ModelRoute{
-		ID:            "route_metrics_codex",
-		ModelName:     model,
-		ProviderID:    provider.ID,
-		ProviderModel: model,
-		Status:        StatusActive,
-	})
-	server := NewWithConfig(store, Config{AdminToken: "dev_admin_token", MetricsEnabled: true})
-	if server.metrics == nil {
-		t.Fatal("expected metrics to be enabled")
-	}
-	server.codexSubscription.Client = &http.Client{Transport: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
-		return upstream(req), nil
-	})}
-	return server, secret
-}
-
-// delayedFirstChunkReader sleeps once before serving its first chunk, simulating
-// an upstream that flushes headers immediately but delays its first SSE event.
-type delayedFirstChunkReader struct {
-	delay   time.Duration
-	body    io.Reader
-	started bool
-}
-
-func (r *delayedFirstChunkReader) Read(p []byte) (int, error) {
-	if !r.started {
-		r.started = true
-		time.Sleep(r.delay)
-	}
-	return r.body.Read(p)
-}
-
-// The Codex gateway measures TTFB on the first SSE event, not on the upstream
-// headers: an upstream that commits the 200 and then waits before its first
-// event must still report roughly its real first-token latency.
-func TestMetricsCodexTimeToFirstByteWaitsForFirstEvent(t *testing.T) {
-	stream := strings.Join([]string{
-		"event: response.output_text.delta",
-		`data: {"type":"response.output_text.delta","delta":"ok"}`,
-		"",
-		"event: response.completed",
-		`data: {"type":"response.completed","response":{"id":"resp_metrics","status":"completed","usage":{"input_tokens":1,"output_tokens":1}}}`,
-		"",
-	}, "\n")
-	server, secret := newMetricsCodexTestServer(t, func(req *http.Request) *http.Response {
-		return &http.Response{
-			StatusCode: http.StatusOK,
-			Header:     http.Header{"Content-Type": []string{"text/event-stream"}},
-			Body:       io.NopCloser(&delayedFirstChunkReader{delay: 300 * time.Millisecond, body: strings.NewReader(stream)}),
-			Request:    req,
-		}
-	})
-	resp := postStream(t, server.Handler(), "/v1/responses", map[string]any{
-		"model":  "metrics-codex-model",
-		"input":  "hi",
-		"stream": true,
-	}, secret)
-	if resp.Code != http.StatusOK {
-		t.Fatalf("responses stream failed: %d %s", resp.Code, resp.Body)
-	}
-
-	body := scrapeMetrics(t, server.Handler())
-	count := findMetricLine(body, "tokenhub_gateway_time_to_first_byte_seconds_count")
-	if count == "" {
-		t.Fatalf("missing Codex TTFB count:\n%s", body)
-	}
-	if got := metricLineValue(t, count); got != 1 {
-		t.Fatalf("expected one Codex TTFB observation, got %v", got)
-	}
-	// 300ms before the first event: the observation must land beyond the 250ms
-	// bucket, not in the near-zero buckets an upstream-header timestamp would hit.
-	bucket := findMetricLine(body, "tokenhub_gateway_time_to_first_byte_seconds_bucket", `le="0.25"`)
-	if bucket == "" {
-		t.Fatalf("missing TTFB bucket:\n%s", body)
-	}
-	if got := metricLineValue(t, bucket); got != 0 {
-		t.Fatalf("TTFB must reflect the delayed first event, got %v observations in le=0.25", got)
-	}
-}
-
-// A Codex stream that ends before any event wrote nothing the client could
-// perceive: it must not count as an interruption and must not record TTFB,
-// matching the chat and Anthropic gateways.
-func TestMetricsCodexEmptyBodyFailureIsNotAnInterruption(t *testing.T) {
-	server, secret := newMetricsCodexTestServer(t, func(req *http.Request) *http.Response {
-		return &http.Response{
-			StatusCode: http.StatusOK,
-			Header:     http.Header{"Content-Type": []string{"text/event-stream"}},
-			Body:       io.NopCloser(strings.NewReader("")),
-			Request:    req,
-		}
-	})
-	resp := postStream(t, server.Handler(), "/v1/responses", map[string]any{
-		"model":  "metrics-codex-model",
-		"input":  "hi",
-		"stream": true,
-	}, secret)
-	// ensureStarted committed the 200 before the body was read, so the client
-	// sees the committed response even though the gateway recorded a 502.
-	if resp.Code != http.StatusOK {
-		t.Fatalf("expected the committed 200, got %d: %s", resp.Code, resp.Body)
-	}
-
-	body := scrapeMetrics(t, server.Handler())
-	if strings.Contains(body, "tokenhub_gateway_stream_interruptions_total{") {
-		t.Fatalf("empty-body Codex failure must not count as interruption:\n%s", body)
-	}
-	if strings.Contains(body, "tokenhub_gateway_time_to_first_byte_seconds{") {
-		t.Fatalf("empty-body Codex failure must not record TTFB:\n%s", body)
-	}
-}
-
-// newMetricsGeminiTestServer mirrors newGeminiCodexTestServer with metrics
-// enabled so the Gemini streaming endpoint can be asserted against.
-func newMetricsGeminiTestServer(t *testing.T, responder func(map[string]any) string) (*Server, string) {
-	t.Helper()
-	store := NewMemoryStore()
-	project := store.CreateProject(Project{Name: "Metrics Gemini Project", Status: StatusActive})
-	_, secret, err := store.CreateAPIKey(project.ID, APIKey{Name: "Metrics Gemini Key", Allowed: []string{geminiCodexTestModel}, Status: StatusActive}, "thk_metrics_gemini")
-	if err != nil {
-		t.Fatal(err)
-	}
-	provider := store.AddProvider(Provider{ID: "prv_gemini_codex", Name: "Gemini Codex", Type: ProviderOpenAICodex, Status: StatusActive, Healthy: true})
-	resource, err := store.AddProviderResource(ProviderResource{
-		ID: "rsrc_gemini_codex", ProviderID: provider.ID, Name: "Gemini Codex Account",
-		ResourceType: ProviderResourceOpenAISubscription, Status: StatusActive, Healthy: true,
-		Options:     codexCapabilityOptionsForTest(geminiCodexTestModel),
-		Credentials: &ProviderResourceCredentials{AccessToken: "gemini-codex-access", AccountID: "gemini-codex-account"},
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	store.AddModel(Model{Name: geminiCodexTestModel, Modality: "chat", ContextWindow: 128000, Status: StatusActive})
-	store.AddRoute(ModelRoute{ID: "route_gemini_codex", ModelName: geminiCodexTestModel, ProviderID: provider.ID, ProviderResourceID: resource.ID, ProviderModel: geminiCodexTestModel, Status: StatusActive})
-	server := NewWithConfig(store, Config{AdminToken: "dev_admin_token", MetricsEnabled: true})
-	if server.metrics == nil {
-		t.Fatal("expected metrics to be enabled")
-	}
-	server.codexSubscription.Client = &http.Client{Transport: roundTripperFunc(func(request *http.Request) (*http.Response, error) {
-		var payload map[string]any
-		if err := json.NewDecoder(request.Body).Decode(&payload); err != nil {
-			t.Fatal(err)
-		}
-		body := responder(payload)
-		return &http.Response{
-			StatusCode: http.StatusOK, Header: http.Header{"Content-Type": []string{"text/event-stream"}},
-			Body: io.NopCloser(strings.NewReader(body)), Request: request,
-		}, nil
-	})}
-	return server, secret
-}
-
-// The Gemini streaming endpoint records TTFB like the chat, Anthropic and Codex
-// gateways: a successful stream is observed exactly once.
-func TestMetricsGeminiStreamRecordsTimeToFirstByteOnce(t *testing.T) {
-	server, secret := newMetricsGeminiTestServer(t, func(request map[string]any) string {
-		return geminiCodexTestSSE(
-			map[string]any{"type": "response.created", "response": map[string]any{"id": "resp_gemini_metrics", "status": "in_progress"}},
-			map[string]any{"type": "response.completed", "response": map[string]any{
-				"id": "resp_gemini_metrics", "status": "completed",
-				"output": []any{map[string]any{"type": "message", "role": "assistant", "content": []any{map[string]any{"type": "output_text", "text": "ok"}}}},
-				"usage":  map[string]any{"input_tokens": 5, "output_tokens": 3, "total_tokens": 8},
-			}},
-		)
-	})
-	response := doGeminiJSON(t, server.Handler(), http.MethodPost, "/v1beta/models/gpt-5.5:streamGenerateContent?alt=sse", map[string]any{
-		"contents": []any{map[string]any{"role": "user", "parts": []any{map[string]any{"text": "hi"}}}},
-	}, secret)
-	if response.Code != http.StatusOK {
-		t.Fatalf("Gemini stream failed: %d %s", response.Code, response.Body)
-	}
-
-	body := scrapeMetrics(t, server.Handler())
-	line := findMetricLine(body, "tokenhub_gateway_time_to_first_byte_seconds_count")
-	if line == "" {
-		t.Fatalf("missing Gemini TTFB count:\n%s", body)
-	}
-	if got := metricLineValue(t, line); got != 1 {
-		t.Fatalf("expected one Gemini TTFB observation, got %v", got)
-	}
-}
-
-// A Gemini stream that fails after its first event counts as one interruption
-// with the upstream error code preserved.
-func TestMetricsGeminiStreamFailureAfterFirstByteCountsInterruption(t *testing.T) {
-	server, secret := newMetricsGeminiTestServer(t, func(request map[string]any) string {
-		// The stream writes a content delta and then ends without
-		// response.completed: the client saw bytes, then the stream broke.
-		return geminiCodexTestSSE(
-			map[string]any{"type": "response.created", "response": map[string]any{"id": "resp_gemini_metrics", "status": "in_progress"}},
-			map[string]any{"type": "response.output_text.delta", "delta": "partial"},
-		)
-	})
-	response := doGeminiJSON(t, server.Handler(), http.MethodPost, "/v1beta/models/gpt-5.5:streamGenerateContent?alt=sse", map[string]any{
-		"contents": []any{map[string]any{"role": "user", "parts": []any{map[string]any{"text": "hi"}}}},
-	}, secret)
-	if response.Code != http.StatusOK {
-		t.Fatalf("a committed Gemini stream still returns the committed 200, got %d", response.Code)
-	}
-
-	body := scrapeMetrics(t, server.Handler())
-	interruption := findMetricLine(body, "tokenhub_gateway_stream_interruptions_total", `provider_id="prv_gemini_codex"`)
-	if interruption == "" {
-		t.Fatalf("missing Gemini interruption series:\n%s", body)
-	}
-	if metricLineValue(t, interruption) != 1 {
-		t.Fatalf("expected one Gemini interruption")
-	}
-	if strings.Contains(interruption, `error_code="none"`) {
-		t.Fatalf("Gemini interruption must carry a non-none error_code:\n%s", interruption)
-	}
-}
-
-// A first write that produces zero bytes — a disconnected client can yield
-// (0, err) — must not be recorded as a first byte: TTFB and interruption both
-// key off firstByteTime, so a zero-byte write failure must leave both empty.
-func TestStreamWriteTrackerZeroByteFirstWriteRecordsNoFirstByte(t *testing.T) {
-	zeroWriter := &zeroByteWrite{err: errors.New("client gone")}
-	tracker := &streamWriteTracker{writer: zeroWriter}
-	n, err := tracker.Write([]byte("event: ping\n\n"))
-	if n != 0 || err == nil {
-		t.Fatalf("expected (0, err) from the failing writer, got (%d, %v)", n, err)
-	}
-	if tracker.WroteData() {
-		t.Fatal("a zero-byte write must not count as data written")
-	}
-	if !tracker.firstByteTime(false).IsZero() {
-		t.Fatal("a failed zero-byte first write must not synthesize a first-byte time")
-	}
-	// A later positive write on the same tracker still records the byte time.
-	tracker.writer = &bytes.Buffer{}
-	start := time.Now()
-	n, err = tracker.Write([]byte("data: ok\n\n"))
-	if n == 0 || err != nil {
-		t.Fatalf("expected a positive write, got (%d, %v)", n, err)
-	}
-	if got := tracker.firstByteTime(false); got.IsZero() || got.Before(start) || got.After(time.Now()) {
-		t.Fatalf("first-byte time must fall within the write window, got %v", got)
-	}
-}
-
-// zeroByteWrite simulates a response writer whose client vanished: it accepts
-// nothing and reports the connection loss.
-type zeroByteWrite struct{ err error }
-
-func (z *zeroByteWrite) Write([]byte) (int, error) { return 0, z.err }
-
-// A native Anthropic stream that forwards an upstream event: error frame after
-// the first byte counts as one interruption: the client saw bytes, then the
-// provider reported a terminal error, and the handler must not record 200.
-func TestMetricsNativeAnthropicErrorAfterFirstByteCountsInterruption(t *testing.T) {
-	upstream := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("content-type", "text/event-stream")
-		_, _ = io.WriteString(w, "event: message_start\n")
-		_, _ = io.WriteString(w, "data: {\"type\":\"message_start\",\"message\":{\"id\":\"msg_m\",\"type\":\"message\",\"role\":\"assistant\",\"content\":[],\"model\":\"upstream-model\",\"stop_reason\":null,\"stop_sequence\":null,\"usage\":{\"input_tokens\":4,\"output_tokens\":0}}}\n\n")
-		_, _ = io.WriteString(w, "event: error\n")
-		_, _ = io.WriteString(w, "data: {\"type\":\"error\",\"error\":{\"type\":\"overloaded_error\",\"message\":\"upstream overloaded\"}}\n\n")
-	}))
-	defer upstream.Close()
-
-	server, secret := newMetricsAnthropicTestServer(t, upstream.URL)
-	resp := doAnthropicRequest(t, server.Handler(), "/v1/messages", map[string]any{
-		"model":      "claude-metrics-test",
-		"max_tokens": 1024,
-		"stream":     true,
-		"messages":   []any{map[string]any{"role": "user", "content": "hi"}},
-	}, "Bearer "+secret, "")
-	if resp.Code != http.StatusOK {
-		t.Fatalf("a committed native stream still returns the committed 200, got %d", resp.Code)
-	}
-	if strings.Count(resp.Body.String(), "event: error") != 1 {
-		t.Fatalf("the terminal error event must reach the client exactly once:\n%s", resp.Body)
-	}
-
-	body := scrapeMetrics(t, server.Handler())
-	interruption := findMetricLine(body, "tokenhub_gateway_stream_interruptions_total", `provider_id="prv_metrics_anthropic"`)
-	if interruption == "" {
-		t.Fatalf("missing native Anthropic interruption series:\n%s", body)
-	}
-	if metricLineValue(t, interruption) != 1 {
-		t.Fatalf("expected one native Anthropic interruption")
-	}
-	if strings.Contains(interruption, `error_code="none"`) {
-		t.Fatalf("native Anthropic interruption must carry a non-none error_code:\n%s", interruption)
-	}
-	if findMetricLine(body, "tokenhub_gateway_time_to_first_byte_seconds_count") == "" {
-		t.Fatalf("TTFB must still be observed for the interrupted stream")
-	}
-}
-
-// newMetricsAnthropicTestServer wires one native Anthropic model with metrics
-// enabled. upstream serves the SSE stream the gateway should forward.
-func newMetricsAnthropicTestServer(t *testing.T, upstreamURL string) (*Server, string) {
-	t.Helper()
-	store := NewMemoryStore()
-	project := store.CreateProject(Project{Name: "metrics-anthropic", Status: StatusActive})
-	_, secret, err := store.CreateAPIKey(project.ID, APIKey{
-		Name:    "metrics-anthropic-key",
-		Allowed: []string{"claude-metrics-test"},
-		Status:  StatusActive,
-	}, "thk_metrics_anthropic")
-	if err != nil {
-		t.Fatal(err)
-	}
-	store.AddModel(Model{
-		Name:          "claude-metrics-test",
-		Family:        "claude",
-		Modality:      "chat",
-		ContextWindow: 200000,
-		Capabilities:  []string{"chat"},
-		Status:        StatusActive,
-	})
-	provider := store.AddProvider(Provider{
-		ID:      "prv_metrics_anthropic",
-		Name:    "Metrics Anthropic",
-		Type:    ProviderAnthropic,
-		BaseURL: upstreamURL,
-		APIKey:  "upstream-secret",
-		Status:  StatusActive,
-		Healthy: true,
-	})
-	store.AddRoute(ModelRoute{
-		ID:            "route_metrics_anthropic",
-		ModelName:     "claude-metrics-test",
-		ProviderID:    provider.ID,
-		ProviderModel: "upstream-model",
-		Priority:      1,
-		Weight:        100,
-		Status:        StatusActive,
-		Strategy:      RouteStrategyPriorityOnly,
-	})
-	return NewWithConfig(store, Config{AdminToken: "dev_admin_token", MetricsEnabled: true}), secret
 }
